fix: move tracing to infrastructure folder
