+++ packages/app/src/infrastructure/infrastructureModule.ts
@@ -2,6 +2,7 @@ import { Logger } from "@byndyusoft/nest-pino";
 import { ApiTags } from "@byndyusoft/nest-swagger";
 import { PromController, PromModule } from "@digikare/nestjs-prom";
 import { Module } from "@nestjs/common";
+import { OpenTelemetryModule } from "nestjs-otel";
 
 import { AboutModule } from "./about/aboutModule";
 import { ClientsModule } from "./clients/clientsModule";
@@ -16,6 +17,7 @@ ApiTags("Infrastructure")(PromController);
 
 @Module({
   imports: [
+    OpenTelemetryModule.forRoot(),
     // Initial modules
     ConfigModule.forRoot(),
     PackageJsonModule,
